requirements.txt
openai==0.28.0
requests==2.32.3
python-docx==1.1.2
